requirements.txt
pillow
tensorflow
numpy
pandas
matplotlip
seaborn
kreas
streamlit
plotly